makes new translation button
